nghttp3_acked_stream_data now takes uint64_t offset
nghttp3_acked_stream_data now takes uint64_t offset because
nghttp3_conn_add_ack_offset takes uint64_t.

+++ lib/includes/nghttp3/nghttp3.h
@@ -1642,7 +1642,7 @@ typedef struct nghttp3_conn nghttp3_conn;
  * :macro:`NGHTTP3_ERR_CALLBACK_FAILURE`.
  */
 typedef int (*nghttp3_acked_stream_data)(nghttp3_conn *conn, int64_t stream_id,
-                                         size_t datalen, void *conn_user_data,
+                                         uint64_t datalen, void *conn_user_data,
                                          void *stream_user_data);
 
 /**
